+++ src/main/java/org/newdawn/slick/geom/TexCoordGenerator.java
@@ -15,5 +15,5 @@ public interface TexCoordGenerator {
 	 * @param y The y coordinate of the vertex being rendered
 	 * @return The texture coordinate to apply
 	 */
-	public Vector2f getCoordFor(float x, float y);
+	public Vector2f getCoordFor(double x, double y);
 }
